backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
google-cloud-storage==2.14.0
firebase-admin==6.4.0
Pillow==11.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
pg8000==1.31.1
cloud-sql-python-connector==1.7.0
pytest==7.4.3
pytest-cov==4.1.0
requests==2.31.0

# Image Processing for Paint-by-Numbers
opencv-python>=4.8.0
scikit-learn>=1.5.0
scipy>=1.13.0
numpy>=2.0.0

# Google Gemini API for image transformation (new API)
google-genai>=0.2.0

